% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_count_matrix.R
\name{samplecounts}
\alias{samplecounts}
\title{Sample mixed or pure count matrix X}
\usage{
samplecounts(
  N,
  M1,
  Ks,
  Kb,
  Nm,
  Na,
  Nh,
  alpha,
  missing = "logit",
  mkappa = 1,
  mtau = 1.5
)
}
\arguments{
\item{N}{An integer. The number of genes.}

\item{M1}{An integer. The number of bulks.}

\item{Ks}{An integer. The number of cell types for pure(single cell) samples.}

\item{Kb}{An integer. The number of cell types for mixed(bulk) samples.}

\item{Nm}{An integer. The number of marker genes in one cell type.}

\item{Na}{An integer. The number of anti-marker genes in one cell type.}

\item{Nh}{An integer. The number of house-keeping genes in one cell type.}

\item{alpha}{A positive vector. Must have a length of \code{K}.This is the concentration parameter of proportion matrix W.}

\item{missing}{A value from 0 to 1 or a character. If missing count rate is fixed, then input 
an value between 0 and 1(included). Another option is \code{"logit"}. This missing rate is based on A and W.}

\item{mkappa}{A scalar. Control the missing rate when \code{missing="logit"}.}

\item{mtau}{A scalar. Control the missing rate when \code{missing="logit"}. The defualt setting of 
\code{mkappa} and \code{mtau} will generate around 0.1 missing in X matrix.}
}
\value{
\code{sample_count_matrix} returns a list of matrices. \code{X} the count matrix.
\code{A} the signature matrix. \code{W} the proportion matrix.
}
\description{
Sample a gene count matrix X(pure,mixed or both) from a signiture matrix A and a proportion matrix W. 
samplecounts<-function(N,M1,Ks,Kb,Nm,Na,Nh,alpha,missing="logit",mkappa=1,mtau=1.5)
}
\examples{
#generate a count matrix X(bulk counts) without single cell counts.
list1<-samplecounts(N=10,M1=5,L=0,Ks=0,Kb=2,Nm=2,Na=0,Nh=5,alpha=c(1,1,1),missing="logit")
X<-list1$X
trueA<-list1$A
trueW<-list1$W
#generate a count matrix X(bulk counts) with single cell counts.
list2<-samplecounts(N=15,M1=5,Ks=3,Kb=3,Nm=2,Na=0,Nh=5,alpha=c(1,2,3),missing="logit")
X<-list2$X
#Check here: the first 5 columns of W are for bulks and the rest of the columns are for single cells.
trueW<-list2$W

}
